feat: Separate cron-related tests
