+++ backend/moment/writing/tests/__init__.py
@@ -1,2 +1,3 @@
 from .tests import *
+from .cron_tests import *
 from .search_tests import *
